requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.50
pydirectinput>=1.0.4
keyboard>=0.13.5
mss>=9.0.0
Pillow>=10.0.0
pyTelegramBotAPI>=4.14.0
psutil>=5.9.0
